+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add 'protontricks'.
 - Add Snapper configuration for '/home' backups/snapshots.
 - Add 16 GB exFAT partition for normal flash drive usage.
+- Add PulseAudio configuration to lower the volume which will helps Mac users.
 
 ## 2.0.0